requirements.txt
fastai==2.5.3
ipykernel==6.25.2
ipywidgets==8.1.1
numpy==1.26.0
pandas==2.1.1
Pillow==10.0.1
scikit-image==0.21.0
seaborn==0.12.2
tqdm==4.66.1
